# openapi.yaml
openapi: 3.1.0
info:
  title: Student Transport System API
  description: Comprehensive API for managing student transportation including buses, routes, trips, bookings, payments, and user management
  version: 1.0.0
  contact:
    name: API Support
    email: support@studenttransport.com
servers:
  - url: http://busmanagementsystem.runasp.net/api
    description: Development server
  - url: https://api.studenttransport.com
    description: Production server
security:
  - BearerAuth: []
paths:
  /api/bookings:
    get:
      summary: Get bookings
      description: Retrieve bookings with optional filtering by tripId, studentId, or date
      parameters:
        - name: tripId
          in: query
          schema:
            type: string
          description: Filter by trip ID
        - name: studentId
          in: query
          schema:
            type: string
          description: Filter by student ID
        - name: date
          in: query
          schema:
            type: string
            format: date
          description: Filter by date
      responses:
        '200':
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
        '401':
          description: Unauthorized - Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      summary: Create a booking
      description: Creates a new booking for a student. Requires active subscription.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tripId:
                  type: string
                  description: ID of the trip to book
                studentId:
                  type: string
                  description: ID of the student making the booking
                stopId:
                  type: string
                  description: ID of the pickup/dropoff stop
              required: [tripId, studentId, stopId]
      responses:
        '201':
          description: Booking created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Subscription required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Trip or student not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: No seats available or duplicate booking
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/settings:
    get:
      summary: Get system settings
      description: Retrieve system configuration settings
      responses:
        '200':
          description: System settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemSettings'
        '401':
          description: Unauthorized - Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    patch:
      summary: Update system settings
      description: Updates system configuration settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SystemSettingsUpdate'
      responses:
        '200':
          description: Settings updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemSettings'
        '401':
          description: Unauthorized - Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Booking:
      type: object
      properties:
        id:
          type: string
          example: "booking-123"
        tripId:
          type: string
          example: "trip-456"
        studentId:
          type: string
          example: "student-789"
        stopId:
          type: string
          example: "stop-123"
        status:
          type: string
          enum: [confirmed, cancelled, completed]
          example: "confirmed"
        date:
          type: string
          format: date
          example: "2025-01-20"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, tripId, studentId, stopId, status, date]
...
    SystemSettings:
      type: object
      properties:
        id:
          type: string
          example: "system-settings"
        systemName:
          type: string
          example: "Bus Management System"
        logo:
          type: string
          example: "/logo.png"
        primaryColor:
          type: string
          example: "#3B82F6"
        secondaryColor:
          type: string
          example: "#10B981"
        maintenanceMode:
          type: boolean
          example: false
        notificationsEnabled:
          type: boolean
          example: true
        updatedAt:
          type: string
          format: date-time
      required: [id, systemName]
...
    SystemSettingsUpdate:
      type: object
      properties:
        systemName:
          type: string
        logo:
          type: string
        primaryColor:
          type: string
        secondaryColor:
          type: string
        maintenanceMode:
          type: boolean
        notificationsEnabled:
          type: boolean
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Internal server error"
        message:
          type: string
          example: "Detailed error message"
      required: [error]
...